Реструктуризирован код: разбит на подпрограммы
